update test command

+++ packages/cli/src/commands/test.ts
@@ -1,10 +1,10 @@
 import path from 'path';
-import babelRequireHook from '@babel/register';
+
 import { getPaths } from '@saruni/internal';
 import execa from 'execa';
 import { run } from 'jest';
-import path from 'path';
 import { CommandBuilder } from 'yargs';
+import babelRequireHook from '@babel/register';
 
 babelRequireHook({
   extends: path.join(getPaths().api.base, '.babelrc.js'),
@@ -19,10 +19,16 @@ export const command = 'test';
 export const desc = 'Runs Jest with the project based setup.';
 
 export const builder: CommandBuilder = (yargs) => {
-  return yargs.option('watchAll', {
-    default: false,
-    type: 'boolean',
-  });
+  return yargs
+    .option('watchAll', {
+      default: false,
+      type: 'boolean',
+    })
+    .option('side', {
+      default: ['api', 'web'],
+      type: 'array',
+      choices: ['api', 'web'],
+    });
 };
 
 export const handler = async (args) => {
@@ -31,19 +37,21 @@ export const handler = async (args) => {
 
     const { db } = require(getPaths().api.db);
 
-    await db.$queryRaw(`DROP SCHEMA IF EXISTS "public" CASCADE`);
+    if (args.side.find('api')) {
+      await db.$queryRaw(`DROP SCHEMA IF EXISTS "public" CASCADE`);
 
-    await db.$disconnect();
+      await db.$disconnect();
 
-    await execa('npx', ['prisma', 'migrate', 'up', '--experimental'], {
-      cwd: getPaths().api.base,
-      env: { DATABASE_URL: process.env.DATABASE_URL_TEST },
-    });
+      await execa('npx', ['prisma', 'migrate', 'up', '--experimental'], {
+        cwd: getPaths().api.base,
+        env: { DATABASE_URL: process.env.DATABASE_URL_TEST },
+      });
 
-    await execa('yarn', ['sr', 'db', 'seed'], {
-      cwd: getPaths().api.base,
-      env: { DATABASE_URL: process.env.DATABASE_URL_TEST },
-    });
+      await execa('yarn', ['sr', 'db', 'seed'], {
+        cwd: getPaths().api.base,
+        env: { DATABASE_URL: process.env.DATABASE_URL_TEST },
+      });
+    }
 
     const testCommand = [
       `--config=${require.resolve('@saruni/config/dist/index.js')}`,
@@ -53,8 +61,12 @@ export const handler = async (args) => {
       testCommand.push('--watchAll');
     }
 
+    if (args.side.length === 1) {
+      testCommand.push(`--projects="<rootDir>/packages/${args.side[0]}"`);
+    }
+
     await run(testCommand);
-  } catch (e) {
-    console.log(e);
+  } catch (error) {
+    console.log(error);
   }
 };
